[bluetooth] Make BluetoothDevice more abstract (#7380)
