+++ bundles/org.openhab.binding.bluetooth/src/main/java/org/openhab/binding/bluetooth/BluetoothDevice.java
@@ -14,18 +14,11 @@ package org.openhab.binding.bluetooth;
 
 import java.time.ZonedDateTime;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.bluetooth.notification.BluetoothConnectionStatusNotification;
-import org.openhab.binding.bluetooth.notification.BluetoothScanNotification;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * The {@link BluetoothDevice} class provides a base implementation of a Bluetooth Low Energy device
@@ -35,8 +28,6 @@ import org.slf4j.LoggerFactory;
  */
 @NonNullByDefault
 public abstract class BluetoothDevice {
-
-    private final Logger logger = LoggerFactory.getLogger(BluetoothDevice.class);
 
     /**
      * Enumeration of Bluetooth connection states
@@ -80,11 +71,6 @@ public abstract class BluetoothDevice {
     }
 
     /**
-     * Current connection state
-     */
-    protected ConnectionState connectionState = ConnectionState.DISCOVERING;
-
-    /**
      * The adapter the device is accessed through
      */
     protected final BluetoothAdapter adapter;
@@ -95,49 +81,6 @@ public abstract class BluetoothDevice {
     protected final BluetoothAddress address;
 
     /**
-     * Manufacturer id
-     */
-    protected @Nullable Integer manufacturer = null;
-
-    /**
-     * Device name.
-     * <p>
-     * Uses the devices long name if known, otherwise the short name if known
-     */
-    protected @Nullable String name;
-
-    protected @Nullable String model;
-    protected @Nullable String serialNumber;
-    protected @Nullable String hardwareRevision;
-    protected @Nullable String firmwareRevision;
-    protected @Nullable String softwareRevision;
-
-    /**
-     * List of supported services
-     */
-    protected final Map<UUID, BluetoothService> supportedServices = new HashMap<>();
-
-    /**
-     * Last known RSSI
-     */
-    protected @Nullable Integer rssi = null;
-
-    /**
-     * Last reported transmitter power
-     */
-    protected @Nullable Integer txPower = null;
-
-    /**
-     * Last time when activity occurred on this device.
-     */
-    protected ZonedDateTime lastSeenTime;
-
-    /**
-     * The event listeners will be notified of device updates
-     */
-    private final List<BluetoothDeviceListener> eventListeners = new CopyOnWriteArrayList<>();
-
-    /**
      * Construct a Bluetooth device taking the Bluetooth address
      *
      * @param adapter
@@ -146,7 +89,6 @@ public abstract class BluetoothDevice {
     public BluetoothDevice(BluetoothAdapter adapter, BluetoothAddress address) {
         this.address = address;
         this.adapter = adapter;
-        this.lastSeenTime = ZonedDateTime.now();
     }
 
     /**
@@ -154,72 +96,42 @@ public abstract class BluetoothDevice {
      *
      * @return The last time this device was active
      */
-    public ZonedDateTime getLastSeenTime() {
-        return lastSeenTime;
-    }
+    public abstract ZonedDateTime getLastSeenTime();
 
     /**
      * Updates the last activity timestamp for this device.
      * Should be called whenever activity occurs on this device.
      *
      */
-    public void updateLastSeenTime() {
-        lastSeenTime = ZonedDateTime.now();
-    }
+    public abstract void updateLastSeenTime();
 
     /**
      * Returns the name of the Bluetooth device.
      *
      * @return The devices name
      */
-    public @Nullable String getName() {
-        return name;
-    }
+    public abstract @Nullable String getName();
 
     /**
-     * Returns the model of the Bluetooth device.
+     * Returns the manufacturer ID of the device
      *
-     * @return The devices model, null if not known
+     * @return an integer with manufacturer ID of the device, or null if not known
      */
-    public @Nullable String getModel() {
-        return model;
-    }
+    public abstract @Nullable Integer getManufacturerId();
 
     /**
-     * Returns the serial number of the Bluetooth device.
+     * Returns the last Transmit Power value or null if no transmit power has been received
      *
-     * @return The serial model, null if not known
+     * @return the last reported transmitter power value in dBm
      */
-    public @Nullable String getSerialNumber() {
-        return serialNumber;
-    }
+    public abstract @Nullable Integer getTxPower();
 
     /**
-     * Returns the hardware revision of the Bluetooth device.
+     * Returns the last Receive Signal Strength Indicator (RSSI) value or null if no RSSI has been received
      *
-     * @return The hardware revision, null if not known
+     * @return the last RSSI value in dBm
      */
-    public @Nullable String getHardwareRevision() {
-        return hardwareRevision;
-    }
-
-    /**
-     * Returns the firmware revision of the Bluetooth device.
-     *
-     * @return The firmware revision, null if not known
-     */
-    public @Nullable String getFirmwareRevision() {
-        return firmwareRevision;
-    }
-
-    /**
-     * Returns the software revision of the Bluetooth device.
-     *
-     * @return The software revision, null if not known
-     */
-    public @Nullable String getSoftwareRevision() {
-        return softwareRevision;
-    }
+    public abstract @Nullable Integer getRssi();
 
     /**
      * Returns the physical address of the device.
@@ -240,134 +152,18 @@ public abstract class BluetoothDevice {
     }
 
     /**
-     * Sets the manufacturer id for the device
-     *
-     * @param manufacturer the manufacturer id
-     */
-    public void setManufacturerId(int manufacturer) {
-        this.manufacturer = manufacturer;
-    }
-
-    /**
-     * Returns the manufacturer ID of the device
-     *
-     * @return an integer with manufacturer ID of the device, or null if not known
-     */
-    public @Nullable Integer getManufacturerId() {
-        return manufacturer;
-    }
-
-    /**
      * Returns a {@link BluetoothService} if the requested service is supported
      *
      * @return the {@link BluetoothService} or null if the service is not supported.
      */
-    public @Nullable BluetoothService getServices(UUID uuid) {
-        return supportedServices.get(uuid);
-    }
+    public abstract @Nullable BluetoothService getServices(UUID uuid);
 
     /**
      * Returns a list of supported service UUIDs
      *
      * @return list of supported {@link BluetoothService}s.
      */
-    public Collection<BluetoothService> getServices() {
-        return supportedServices.values();
-    }
-
-    /**
-     * Sets the device transmit power
-     *
-     * @param power the current transmitter power in dBm
-     */
-    public void setTxPower(int txPower) {
-        this.txPower = txPower;
-    }
-
-    /**
-     * Returns the last Transmit Power value or null if no transmit power has been received
-     *
-     * @return the last reported transmitter power value in dBm
-     */
-    public @Nullable Integer getTxPower() {
-        return txPower;
-    }
-
-    /**
-     * Sets the current Receive Signal Strength Indicator (RSSI) value
-     *
-     * @param rssi the current RSSI value in dBm
-     * @return true if the RSSI has changed, false if it was the same as previous
-     */
-    public boolean setRssi(int rssi) {
-        boolean changed = (this.rssi == null || this.rssi != rssi);
-        this.rssi = rssi;
-
-        return changed;
-    }
-
-    /**
-     * Returns the last Receive Signal Strength Indicator (RSSI) value or null if no RSSI has been received
-     *
-     * @return the last RSSI value in dBm
-     */
-    public @Nullable Integer getRssi() {
-        return rssi;
-    }
-
-    /**
-     * Set the name of the device
-     *
-     * @param name a {@link String} defining the device name
-     */
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    /**
-     * Set the model of the device
-     *
-     * @param model a {@link String} defining the device model
-     */
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    /**
-     * Set the serial number of the device
-     *
-     * @param model a {@link String} defining the serial number
-     */
-    public void setSerialNumberl(String serialNumber) {
-        this.serialNumber = serialNumber;
-    }
-
-    /**
-     * Set the hardware revision of the device
-     *
-     * @param model a {@link String} defining the hardware revision
-     */
-    public void setHardwareRevision(String hardwareRevision) {
-        this.hardwareRevision = hardwareRevision;
-    }
-
-    /**
-     * Set the firmware revision of the device
-     *
-     * @param model a {@link String} defining the firmware revision
-     */
-    public void setFirmwareRevision(String firmwareRevision) {
-        this.firmwareRevision = firmwareRevision;
-    }
-
-    /**
-     * Set the software revision of the device
-     *
-     * @param model a {@link String} defining the software revision
-     */
-    public void setSoftwareRevision(String softwareRevision) {
-        this.softwareRevision = softwareRevision;
-    }
+    public abstract Collection<BluetoothService> getServices();
 
     /**
      * Check if the device supports the specified service
@@ -375,18 +171,14 @@ public abstract class BluetoothDevice {
      * @param uuid the service {@link UUID}
      * @return true if the service is supported
      */
-    public boolean supportsService(UUID uuid) {
-        return supportedServices.containsKey(uuid);
-    }
+    public abstract boolean supportsService(UUID uuid);
 
     /**
      * Get the current connection state for this device
      *
      * @return the current {@link ConnectionState}
      */
-    public ConnectionState getConnectionState() {
-        return connectionState;
-    }
+    public abstract ConnectionState getConnectionState();
 
     /**
      * Connects to a device. This is an asynchronous method. Once the connection state is updated, the
@@ -396,9 +188,7 @@ public abstract class BluetoothDevice {
      *
      * @return true if the connection process is started successfully
      */
-    public boolean connect() {
-        return false;
-    }
+    public abstract boolean connect();
 
     /**
      * Disconnects from a device. Once the connection state is updated, the
@@ -409,9 +199,7 @@ public abstract class BluetoothDevice {
      *
      * @return true if the disconnection process is started successfully
      */
-    public boolean disconnect() {
-        return false;
-    }
+    public abstract boolean disconnect();
 
     /**
      * Starts a discovery on a device. This will iterate through all services and characteristics to build up a view of
@@ -421,9 +209,7 @@ public abstract class BluetoothDevice {
      *
      * @return true if the discovery process is started successfully
      */
-    public boolean discoverServices() {
-        return false;
-    }
+    public abstract boolean discoverServices();
 
     /**
      * Gets a Bluetooth characteristic if it is known.
@@ -436,7 +222,7 @@ public abstract class BluetoothDevice {
      * @return the {@link BluetoothCharacteristic} or null if the characteristic is not found in the device
      */
     public @Nullable BluetoothCharacteristic getCharacteristic(UUID uuid) {
-        for (BluetoothService service : supportedServices.values()) {
+        for (BluetoothService service : getServices()) {
             if (service.providesCharacteristic(uuid)) {
                 return service.getCharacteristic(uuid);
             }
@@ -457,9 +243,7 @@ public abstract class BluetoothDevice {
      * @param characteristic the {@link BluetoothCharacteristic} to read.
      * @return true if the characteristic read is started successfully
      */
-    public boolean readCharacteristic(BluetoothCharacteristic characteristic) {
-        return false;
-    }
+    public abstract boolean readCharacteristic(BluetoothCharacteristic characteristic);
 
     /**
      * Writes a characteristic. Only a single read or write operation can be requested at once. Attempting to perform an
@@ -471,9 +255,7 @@ public abstract class BluetoothDevice {
      * @param characteristic the {@link BluetoothCharacteristic} to read.
      * @return true if the characteristic write is started successfully
      */
-    public boolean writeCharacteristic(BluetoothCharacteristic characteristic) {
-        return false;
-    }
+    public abstract boolean writeCharacteristic(BluetoothCharacteristic characteristic);
 
     /**
      * Enables notifications for a characteristic. Only a single read or write operation can be requested at once.
@@ -485,9 +267,7 @@ public abstract class BluetoothDevice {
      * @param characteristic the {@link BluetoothCharacteristic} to receive notifications for.
      * @return true if the characteristic notification is started successfully
      */
-    public boolean enableNotifications(BluetoothCharacteristic characteristic) {
-        return false;
-    }
+    public abstract boolean enableNotifications(BluetoothCharacteristic characteristic);
 
     /**
      * Disables notifications for a characteristic. Only a single read or write operation can be requested at once.
@@ -497,9 +277,7 @@ public abstract class BluetoothDevice {
      * @param characteristic the {@link BluetoothCharacteristic} to disable notifications for.
      * @return true if the characteristic notification is stopped successfully
      */
-    public boolean disableNotifications(BluetoothCharacteristic characteristic) {
-        return false;
-    }
+    public abstract boolean disableNotifications(BluetoothCharacteristic characteristic);
 
     /**
      * Enables notifications for a descriptor. Only a single read or write operation can be requested at once.
@@ -511,9 +289,7 @@ public abstract class BluetoothDevice {
      * @param descriptor the {@link BluetoothDescriptor} to receive notifications for.
      * @return true if the descriptor notification is started successfully
      */
-    public boolean enableNotifications(BluetoothDescriptor descriptor) {
-        return false;
-    }
+    public abstract boolean enableNotifications(BluetoothDescriptor descriptor);
 
     /**
      * Disables notifications for a descriptor. Only a single read or write operation can be requested at once.
@@ -523,9 +299,7 @@ public abstract class BluetoothDevice {
      * @param descriptor the {@link BluetoothDescriptor} to disable notifications for.
      * @return true if the descriptor notification is stopped successfully
      */
-    public boolean disableNotifications(BluetoothDescriptor descriptor) {
-        return false;
-    }
+    public abstract boolean disableNotifications(BluetoothDescriptor descriptor);
 
     /**
      * Adds a service to the device.
@@ -533,14 +307,7 @@ public abstract class BluetoothDevice {
      * @param service the new {@link BluetoothService} to add
      * @return true if the service was added or false if the service was already supported
      */
-    protected boolean addService(BluetoothService service) {
-        if (supportedServices.containsKey(service.getUuid())) {
-            return false;
-        }
-        logger.trace("Adding new service to device {}: {}", address, service);
-        supportedServices.put(service.getUuid(), service);
-        return true;
-    }
+    protected abstract boolean addService(BluetoothService service);
 
     /**
      * Adds a list of services to the device
@@ -567,11 +334,9 @@ public abstract class BluetoothDevice {
      * @return the {@link BluetoothService} or null if the service was not found
      */
     protected @Nullable BluetoothService getServiceByHandle(int handle) {
-        synchronized (supportedServices) {
-            for (BluetoothService service : supportedServices.values()) {
-                if (service.getHandleStart() <= handle && service.getHandleEnd() >= handle) {
-                    return service;
-                }
+        for (BluetoothService service : getServices()) {
+            if (service.getHandleStart() <= handle && service.getHandleEnd() >= handle) {
+                return service;
             }
         }
         return null;
@@ -588,7 +353,6 @@ public abstract class BluetoothDevice {
         if (service != null) {
             return service.getCharacteristicByHandle(handle);
         }
-
         return null;
     }
 
@@ -597,34 +361,27 @@ public abstract class BluetoothDevice {
      *
      * @param listener the {@link BluetoothDeviceListener} to add
      */
-    public void addListener(BluetoothDeviceListener listener) {
-        eventListeners.add(listener);
-    }
+    public abstract void addListener(BluetoothDeviceListener listener);
 
     /**
      * Removes a device listener
      *
      * @param listener the {@link BluetoothDeviceListener} to remove
      */
-    public void removeListener(BluetoothDeviceListener listener) {
-        eventListeners.remove(listener);
-    }
+    public abstract void removeListener(BluetoothDeviceListener listener);
 
     /**
      * Checks if this device has any listeners
      *
      * @return true if this device has listeners
      */
-    public boolean hasListeners() {
-        return !eventListeners.isEmpty();
-    }
+    public abstract boolean hasListeners();
 
     /**
      * Releases resources that this device is using.
      *
      */
-    protected void dispose() {
-    }
+    protected abstract void dispose();
 
     /**
      * Notify the listeners of an event
@@ -632,67 +389,6 @@ public abstract class BluetoothDevice {
      * @param event the {@link BluetoothEventType} of this event
      * @param args an array of arguments to pass to the callback
      */
-    protected void notifyListeners(BluetoothEventType event, Object... args) {
-        for (BluetoothDeviceListener listener : eventListeners) {
-            try {
-                switch (event) {
-                    case SCAN_RECORD:
-                        listener.onScanRecordReceived((BluetoothScanNotification) args[0]);
-                        break;
-                    case CONNECTION_STATE:
-                        listener.onConnectionStateChange((BluetoothConnectionStatusNotification) args[0]);
-                        break;
-                    case SERVICES_DISCOVERED:
-                        listener.onServicesDiscovered();
-                        break;
-                    case CHARACTERISTIC_READ_COMPLETE:
-                        listener.onCharacteristicReadComplete((BluetoothCharacteristic) args[0],
-                                (BluetoothCompletionStatus) args[1]);
-                        break;
-                    case CHARACTERISTIC_WRITE_COMPLETE:
-                        listener.onCharacteristicWriteComplete((BluetoothCharacteristic) args[0],
-                                (BluetoothCompletionStatus) args[1]);
-                        break;
-                    case CHARACTERISTIC_UPDATED:
-                        listener.onCharacteristicUpdate((BluetoothCharacteristic) args[0]);
-                        break;
-                    case DESCRIPTOR_UPDATED:
-                        listener.onDescriptorUpdate((BluetoothDescriptor) args[0]);
-                        break;
-                }
-            } catch (Exception e) {
-                logger.error("Failed to inform listener '{}': {}", listener, e.getMessage(), e);
-            }
-        }
-    }
+    protected abstract void notifyListeners(BluetoothEventType event, Object... args);
 
-    @Override
-    public String toString() {
-        StringBuilder builder = new StringBuilder();
-        builder.append("BluetoothDevice [address=");
-        builder.append(address);
-        builder.append(", manufacturer=");
-        builder.append(manufacturer);
-        if (BluetoothCompanyIdentifiers.get(manufacturer) != null) {
-            builder.append(" (");
-            builder.append(BluetoothCompanyIdentifiers.get(manufacturer));
-            builder.append(')');
-        }
-        builder.append(", name=");
-        builder.append(name);
-        builder.append(", model=");
-        builder.append(model);
-        builder.append(", serialNumber=");
-        builder.append(serialNumber);
-        builder.append(", hardwareRevision=");
-        builder.append(hardwareRevision);
-        builder.append(", firmwareRevision=");
-        builder.append(firmwareRevision);
-        builder.append(", softwareRevision=");
-        builder.append(softwareRevision);
-        builder.append(", rssi=");
-        builder.append(rssi);
-        builder.append(']');
-        return builder.toString();
-    }
 }
