+++ bundles/org.openhab.binding.bluetooth/src/main/java/org/openhab/binding/bluetooth/discovery/internal/BluetoothDiscoveryService.java
@@ -269,7 +269,7 @@ public class BluetoothDiscoveryService extends AbstractDiscoveryService implemen
             discoveryFutures.put(adapter, new SnapshotFuture(snapshot, future));
         }
 
-        private CompletableFuture<DiscoveryResult> startDiscoveryProcess(BluetoothDevice device) {
+        private CompletableFuture<DiscoveryResult> startDiscoveryProcess(BluetoothDeviceSnapshot device) {
             return CompletableFuture.supplyAsync(new BluetoothDiscoveryProcess(device, participants, adapters),
                     scheduler);
         }
