+++ bundles/org.openhab.binding.bluetooth/src/test/java/org/openhab/binding/bluetooth/MockBluetoothDevice.java
@@ -15,6 +15,7 @@ package org.openhab.binding.bluetooth;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import org.eclipse.jdt.annotation.NonNull;
 import org.openhab.binding.bluetooth.BluetoothCharacteristic.GattCharacteristic;
 import org.openhab.binding.bluetooth.notification.BluetoothConnectionStatusNotification;
 
@@ -23,7 +24,7 @@ import org.openhab.binding.bluetooth.notification.BluetoothConnectionStatusNotif
  *
  * @author Connor Petty - Initial contribution
  */
-public class MockBluetoothDevice extends BluetoothDevice {
+public class MockBluetoothDevice extends BaseBluetoothDevice {
 
     private AtomicBoolean servicesDiscovered = new AtomicBoolean(false);
 
@@ -88,4 +89,30 @@ public class MockBluetoothDevice extends BluetoothDevice {
     protected void notifyListeners(BluetoothEventType event, Object... args) {
         CompletableFuture.runAsync(() -> super.notifyListeners(event, args));
     }
+
+    @Override
+    public boolean writeCharacteristic(@NonNull BluetoothCharacteristic characteristic) {
+        return false;
+    }
+
+    @Override
+    public boolean enableNotifications(@NonNull BluetoothCharacteristic characteristic) {
+        return false;
+    }
+
+    @Override
+    public boolean disableNotifications(@NonNull BluetoothCharacteristic characteristic) {
+        return false;
+    }
+
+    @Override
+    public boolean enableNotifications(@NonNull BluetoothDescriptor descriptor) {
+        return false;
+    }
+
+    @Override
+    public boolean disableNotifications(@NonNull BluetoothDescriptor descriptor) {
+        return false;
+    }
+
 }
